// app/helpers/interfaces.tsx
export interface CandleStickProps {
  t: number;
  o: string;
  c: string;
  h: string;
  l: string;
}
export interface WAGMIStickProps {
  timestamp: number;
  open: number;
  high: number;
  low: number;
  close: number;
}
export interface LineChartProps {
  E: number;
  e: string;
  p: string;
  s: string;
}
export interface WAGMILineChartsProps {
  timestamp: number;
  value: number;
}